Client:  -Re-iterated how client waits for responses and send requests. Now the clients starts 1 thread dedicated to listening for responses. This allows the server to send more than one response at a time.
 Server:
 -Implemented the StopOrder & CancelOrder
 -Implemented automatic logout if client is inactive for more then X seconds( where X = server.config->client_max_timeout).

+++ src/Messages/Requests/MarketOrderRequest.java
@@ -6,7 +6,7 @@ import Messages.OperationType;
 public class MarketOrderRequest extends Request
 {
 	String type;
-	int size;
+	long size;
 	String owner;
 
 	//Needed 4 correct gson deserialization (when using gson.fromJson(..., ...) this constructor gets used, not the one below).
@@ -15,7 +15,7 @@ public class MarketOrderRequest extends Request
 		super(OperationType.MARKET_ORDER);
 	}
 
-	public MarketOrderRequest(OrderKind kind, int size, String owner)
+	public MarketOrderRequest(OrderKind kind, long size, String owner)
 	{
 		this();
 		this.type = kind.GetName();
@@ -28,7 +28,7 @@ public class MarketOrderRequest extends Request
 		return OrderKind.Get(type);
 	}
 
-	public int GetSize()
+	public long GetSize()
 	{
 		return size;
 	}
